+++ asset_allocation/return_calc/return_metrics.py
@@ -67,6 +67,7 @@ class ReturnMetrics(object):
         self,
         allocation: np.ndarray,
         this_returns: np.ndarray,
+        stopped_securities_matrix: np.ndarray,
         indexes: np.ndarray,
         next_allocation: np.ndarray = None,
         trans_cost: float = 0.0,
@@ -82,6 +83,8 @@ class ReturnMetrics(object):
             current allocation
         this_returns: np.array
             forecasted returns per asset
+        stopped_securities_matrix: np.array
+            matrix of stopped securities per trading period
         indexes: np.array
             index column for returned DataFrame
             should be set to date range
@@ -96,6 +99,12 @@ class ReturnMetrics(object):
             return: float
         """
         n_obs = len(this_returns)
+        traded_m = np.bitwise_not(stopped_securities_matrix)
+        first_stop = (
+            stopped_securities_matrix.cumsum(axis=0).cumsum(axis=0) == 1
+        ) * trans_cost
+
+        this_returns = this_returns * traded_m
         cumulative_returns = np.cumprod(this_returns + 1, axis=0)
         cumulative_returns = np.where(
             np.isnan(cumulative_returns), 0, cumulative_returns
@@ -105,6 +114,7 @@ class ReturnMetrics(object):
         ending_allocation = (
             ending_allocation.T / np.nansum(ending_allocation, axis=1)
         ).T
+        stopped_ending = ending_allocation * traded_m
 
         actual_returns = allocation @ cumulative_returns.T
         actual_returns = np.insert(actual_returns, 0, 1)
@@ -115,8 +125,9 @@ class ReturnMetrics(object):
         next_allocation_m[-1] = next_allocation
         trans_cost_m = np.zeros((n_obs, self.universe_size))
         trans_cost_m[-1] = trans_cost
+        trans_cost_m = np.max([trans_cost_m, first_stop], axis=0)
         if next_allocation is not None:
-            turnover = abs(next_allocation_m - ending_allocation)
+            turnover = abs(next_allocation_m - stopped_ending)
             this_trans_cost = (turnover * trans_cost_m).sum(axis=1)
             actual_returns -= this_trans_cost