+++ asset_allocation/strategies/pick_all.py
@@ -52,6 +52,9 @@ class PickAll(strategy.Strategy):
         self.portfolio_return_engine.assign(
             date=date, price_return=price_return, annualize_factor=annualize_factor
         )
+        self.stop_loss.assign(
+            date=date, price_return=price_return, annualize_factor=annualize_factor
+        )
         # only calculate cov matrix on rebalance dates to save time
         if date in self.rebalance_dates:
             self.risk_engine.assign(